Task Reminders Feature
Added a new field to the task data structure to store the reminder interval.
Updated the task creation and editing functions to allow users to set the reminder interval of a task.
Created a function that checks if the current time is within the reminder interval for any tasks. If it is, a reminder notification should be sent.
Run the reminder checking function in the background at regular intervals.